Getting rid of code containing CommandLine().
Using the CommandLine( ) was my bad. I took the outline of main() here from one of my labs. Its original function was to handle any arguments specified by the user, such as only returning sequences of a specific length. We may find use for it, but with the current code as it is, FastAreader() handles stdin.